pyproject.toml
# ============================================
# AUTOTAREFAS - Configuração do Projeto
# ============================================

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# === METADADOS DO PROJETO ===
[project]
name = "autotarefas"
version = "0.1.0"
description = "Sistema de automação de tarefas repetitivas do computador"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
authors = [
    {name = "AutoTarefas Team", email = ""},
]
keywords = [
    "automation",
    "backup",
    "cleaner",
    "monitor",
    "scheduler",
    "cli",
    "tasks",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: System :: Archiving :: Backup",
    "Topic :: System :: Monitoring",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
]

# === DEPENDÊNCIAS DE PRODUÇÃO ===
# Versões verificadas em 31/Dez/2025
dependencies = [
    "click>=8.1.0,<9.0.0",           # CLI framework (testado: 8.3.1)
    "rich>=13.0.0,<15.0.0",          # Terminal UI (testado: 14.2.0)
    "loguru>=0.7.0,<1.0.0",          # Logging (testado: 0.7.3)
    "schedule>=1.2.0,<2.0.0",        # Agendamento (testado: 1.2.2)
    "psutil>=5.9.0,<8.0.0",          # Monitoramento (testado: 7.2.1)
    "python-dotenv>=1.0.0,<2.0.0",   # Config (testado: 1.2.1)
]

# === DEPENDÊNCIAS OPCIONAIS ===
[project.optional-dependencies]
# Para relatórios com Excel
reports = [
    "pandas>=2.0.0,<3.0.0",          # (testado: 2.3.3)
    "openpyxl>=3.1.0,<4.0.0",
]

# Para templates de email
email = [
    "jinja2>=3.1.0,<4.0.0",
]

# Todas as dependências opcionais
all = [
    "autotarefas[reports,email]",
]

# Para desenvolvimento
dev = [
    "pytest>=8.0.0,<10.0.0",         # (testado: 9.0.2)
    "pytest-cov>=4.1.0,<8.0.0",      # (testado: 7.0.0)
    "pytest-mock>=3.11.0,<4.0.0",    # (testado: 3.15.1)
    "ruff>=0.1.0,<1.0.0",
    "mypy>=1.5.0,<2.0.0",
    "pre-commit>=3.4.0,<4.0.0",
]

# Para documentação
docs = [
    "mkdocs>=1.5.0,<2.0.0",          # (testado: 1.6.1)
    "mkdocs-material>=9.4.0,<10.0.0",
]

# === URLs DO PROJETO ===
[project.urls]
Homepage = "https://github.com/paulor007/autotarefas"
Documentation = "https://paulor007.github.io/autotarefas"
Repository = "https://github.com/paulor007/autotarefas"
Issues = "https://github.com/paulor007/autotarefas/issues"
Changelog = "https://github.com/paulor007/autotarefas/blob/main/CHANGELOG.md"

# === PONTO DE ENTRADA CLI ===
[project.scripts]
autotarefas = "autotarefas.cli.main:cli"

# === CONFIGURAÇÃO DO HATCH ===
[tool.hatch.build.targets.wheel]
packages = ["src/autotarefas"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/README.md",
    "/LICENSE",
    "/CHANGELOG.md",
]

# === CONFIGURAÇÃO DO RUFF (Linter + Formatter) ===
[tool.ruff]
target-version = "py312"
line-length = 120

src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B905",   # zip without explicit strict
]

[tool.ruff.lint.isort]
known-first-party = ["autotarefas"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*" = ["ARG001"]  # Unused function args (fixtures)

# === CONFIGURAÇÃO DO MYPY (Type Checking) ===
[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
namespace_packages = true
explicit_package_bases = true

[[tool.mypy.overrides]]
module = [
    "schedule.*",
    "loguru.*",
]
ignore_missing_imports = true

# === CONFIGURAÇÃO DO PYTEST ===
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
python_classes = ["Test*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "-ra",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "e2e: marks tests as end-to-end tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

# === CONFIGURAÇÃO DO COVERAGE ===
[tool.coverage.run]
source = ["src/autotarefas"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "if typing.TYPE_CHECKING:",
]
fail_under = 80
show_missing = true
